Added caching in financial data api
