+++ src/app/api/Fundamentals/route.js
@@ -2,32 +2,8 @@ import { NextResponse } from "next/server";
 import axios from "axios";
 import { XMLParser } from "fast-xml-parser";
 import { NseIndia } from "stock-nse-india";
+import { getCachedData, setCachedData } from "./CachedFinancialData";
 const nseIndia = new NseIndia();
-
-const financialDataCache = new Map();
-
-export function getCachedData(symbol) {
-  const entry = financialDataCache.get(symbol);
-  if (!entry) return null;
-
-  const now = Date.now();
-  const { data, timestamp, ttl } = entry;
-
-  if (now - timestamp < ttl) {
-    return data;
-  } else {
-    financialDataCache.delete(symbol);
-    return null;
-  }
-}
-
-export function setCachedData(symbol, data, ttl = 1000 * 60 * 15) {
-  financialDataCache.set(symbol, {
-    data,
-    timestamp: Date.now(),
-    ttl,
-  });
-}
 
 function getValue(field, contextRef = null) {
   if (!field) return "Not Found";
